Don't send request to the server if assignee wasn't changed
When the user clicks on the 'Assign' button, but the assignee wasn't
changed there is no need to send a request to the server.

Sending a request to the server when the assignee wasn't changed can
lead to confusing error messages, e.g. if the assignee field is empty
and the user clicks on the 'Assign' button, it could result in the
error 'Delete Assignee not permitted'. This is confusing because there
is no assignee which could be deleted.

+++ gerrit-gwtui/src/main/java/com/google/gerrit/client/change/Assignee.java
@@ -63,6 +63,7 @@ public class Assignee extends Composite {
 
   private AssigneeSuggestOracle assigneeSuggestOracle;
   private Change.Id changeId;
+  private String currentAssignee;
 
   Assignee() {
     assigneeSuggestOracle = new AssigneeSuggestOracle();
@@ -93,7 +94,8 @@ public class Assignee extends Composite {
 
   void set(ChangeInfo info) {
     this.changeId = info.legacyId();
-    assigneeLink.setText(info.assignee() != null ? info.assignee().name() : "");
+    this.currentAssignee = info.assignee() != null ? info.assignee().name() : "";
+    assigneeLink.setText(currentAssignee);
     assigneeSuggestOracle.setChange(changeId);
     editAssigneeIcon.setVisible(Gerrit.isSignedIn());
   }
@@ -124,12 +126,17 @@ public class Assignee extends Composite {
   }
 
   private void editAssignee(final String assignee) {
+    if (currentAssignee.equals(assignee)) {
+      return;
+    }
+
     if (assignee.isEmpty()) {
       ChangeApi.deleteAssignee(changeId.get(),
           new GerritCallback<AccountInfo>() {
             @Override
             public void onSuccess(AccountInfo result) {
               onCloseForm();
+              currentAssignee = "";
               assigneeLink.setText("");
             }
 
@@ -151,6 +158,7 @@ public class Assignee extends Composite {
             @Override
             public void onSuccess(AccountInfo result) {
               onCloseForm();
+              currentAssignee = result.name();
               assigneeLink.setText(result.name());
             }
